+++ src/commands/manage-favorites/favorites-view-message.ts
@@ -1,29 +1,37 @@
-import type { FavoritesNode } from "../../config-interface";
+import type { ConfigTypeMap, FavoritesNode } from "../../config-interface";
 import type { FavoritesSectionType } from "./favorites-section-type";
+import type { Message } from "./utility/message-bus";
 
 /** Message sent by the favorites manager view. */
 export type FavoritesViewMessage =
 	GetFavorites |
 	GetUnicodeData |
+	GetConfigValue |
 	ChangeEvent |
 	SaveEvent;
 
-export interface GetFavorites
+export interface GetFavorites extends Message
 {
 	type: 'get-favorites';
 }
 
-export interface GetUnicodeData
+export interface GetUnicodeData extends Message
 {
 	type: 'get-unicode-data';
 }
 
-export interface ChangeEvent
+export interface GetConfigValue extends Message
+{
+	type: 'get-config-value';
+	config: keyof ConfigTypeMap;
+}
+
+export interface ChangeEvent extends Message
 {
 	type: 'changed';
 }
 
-export interface SaveEvent
+export interface SaveEvent extends Message
 {
 	type: 'save';
 	sections: SaveEventSection[];