+++ src/commands/manage-favorites/utility/types.ts
@@ -1,0 +1,1 @@
+export type Await<T> = T extends Promise<infer R> ? R : T;