+++ src/commands/manage-favorites/utility/unicode-data.ts
@@ -2,7 +2,7 @@ import { messageBus } from './message-bus';
 
 export const asyncData = (async () =>
 {
-	const response = messageBus.call('unicode-data', { type: 'get-unicode-data' });
+	const response = messageBus.call<'unicode-data'>({ type: 'get-unicode-data' });
 
 	return (await response).data;
 })();